[#10] Replace getLangLong function & Put styles in seprate file
