+++ components/switchCampuses/styles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  btn: {
+    bottom: '97%',
+    width: '30%',
+    marginHorizontal: 5,
+  },
+  container: {
+    flexDirection: 'row',
+    position: 'absolute',
+    justifyContent: 'center'
+  },
+});
+
+export default styles;
